requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.8
django-background-tasks==1.2.5
django-celery-beat==2.5.0
django-compat==1.0.15
django-crontab==0.7.1
django-phonenumber-field==7.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
flower==2.0.1
google-auth==2.25.1
humanize==4.9.0
idna==3.6
importlib-metadata==4.13.0
kombu==5.3.4
mysqlclient==2.2.0
phonenumbers==8.13.26
Pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.6
vine==5.1.0
wcwidth==0.2.12
zipp==3.17.0
gunicorn==20.1.0
django-cors-headers==4.3.0
tensorflow>=2.2.0;sys_platform != 'darwin'
tensorflow_macos>=2.5.0;sys_platform == 'darwin'
tensorflow-hub==0.12.0
pillow
numpy